Contains the Solution file & the Notes C# Folder
The Notes.Android file was "too large" for Github's liking. Pulling the contents out from within is possible, but it has too many files with a similar name so I'll be uploading it to a sub branch to keep them separated to prevent accidentally crossing the files to one another. The parent folder for each, Notes and Notes.Android will likely need to be created to keep the directory path clean if you download.

The path should be Notes>Folder Containing Solution + Notes Folder> Notes + Notes.Android Folder> Respective files for each folder.