Contains all but Obj Folder
Github refuses to accept the Obj folder as a single file due to it containing more than 100 files/folders within it. As such, it is not included in this branch. For now, until a better solution presents itself, I'll create a sub branch of this branch and put the obj files in there individually to get around Githubs failings.